+++ src/containers/restaurant_post_review.jsx
@@ -85,7 +85,17 @@ class PostReview extends Component {
 	renderReview(){
 		var rating = parseInt(this.state.rating);	
 		// to convert date from : http://stackoverflow.com/questions/13459866/javascript-change-date-into-format-of-dd-mm-yyyy
-		var date = Date();
+		var date = function js_yyyy_mm_dd_hh_mm_ss () {
+			  var now = new Date();
+			  var year = "" + now.getFullYear();
+			  var month = "" + (now.getMonth() + 1); if (month.length == 1) { month = "0" + month; }
+			  var day = "" + now.getDate(); if (day.length == 1) { day = "0" + day; }
+			  var hour = "" + now.getHours(); if (hour.length == 1) { hour = "0" + hour; }
+			  var minute = "" + now.getMinutes(); if (minute.length == 1) { minute = "0" + minute; }
+			  var second = "" + now.getSeconds(); if (second.length == 1) { second = "0" + second; }
+			  return year + "-" + month + "-" + day + " " + hour + ":" + minute + ":" + second;
+			}();
+
 		return (
 			<div >
 				<Reviews image={this.state.gender} name={this.state.name} date={date} rating={rating} comment={this.state.comment} />
@@ -94,11 +104,11 @@ class PostReview extends Component {
 	}
 	postReview(event){
 		event.preventDefault();
-		console.log("Current state name:" + this.state.name);
-		console.log("Current state rating:" + this.state.rating);
-		console.log("Current state rating:" + this.state.gender);
-		console.log("Current state comment:" + this.state.comment);
-		console.log("Current state date:" + this.state.date);
+		// console.log("Current state name:" + this.state.name);
+		// console.log("Current state rating:" + this.state.rating);
+		// console.log("Current state rating:" + this.state.gender);
+		// console.log("Current state comment:" + this.state.comment);
+		// console.log("Current state date:" + this.state.date);
 		var reviewContainer = document.querySelector('#reviewPost');
 		var ratingPicker = document.querySelector('#ratingPicker');
 		reviewPost.classList.remove('hidden');