+++ src/containers/restaurant_post_review.jsx
@@ -6,20 +6,12 @@
 
 
 import React, { Component } from 'react';
-
-// imports the redux library, to forge a connection between the state and the main application
-// imports the connect function
 import { connect } from 'react-redux';
-
-// imports the function within react to bind action creators to components
 import { bindActionCreators } from 'redux';
-
-// imports action creators
 import postReview from '../actions/post_review';
 
 import Reviews from '../components/reviews';
 
-// THESE CLASSES MUST HAVE A CAPITAL AT THE BEGINNING OR THE CLASS WILL NOT WORK
 function setState(stateBinding){
 	this.setState({
 			name: state.target.value,
@@ -27,21 +19,17 @@ function setState(stateBinding){
 	console.log(this.state.name);
 };
 class PostReview extends Component {
-	// sets up the state handler for which books to display
+
 	constructor(props){
 		super(props);
 		
 		this.state = {
-
-			// sets the post data
 			name: '',
 			rating: '2',
 			gender: '',
-			comment: '',
-			date: ''
+			comment: ''
 		};
-		
-		// binds the search input
+
 		this.postName = this.postName.bind(this);
 		this.postRating = this.postRating.bind(this);
 		this.postReview = this.postReview.bind(this);
@@ -54,7 +42,7 @@ class PostReview extends Component {
 		this.setState({
 			name: name.target.value,
 		});
-		console.log(this.state.name);
+		//console.log(this.state.name);
 	}
 	postRating(rating){
 
@@ -62,21 +50,21 @@ class PostReview extends Component {
 			rating: rating.target.value,
 
 		});
-		console.log(this.state.rating);
+		//console.log(this.state.rating);
 	}
 	postComment(comment){
 
 		this.setState({
 			comment: comment.target.value
 		});
-		console.log(this.state.comment);
+		//console.log(this.state.comment);
 	}
 	postGender(gender){
 
 		this.setState({
 			gender: gender.target.value
 		});
-		console.log(this.state.gender);
+		//console.log(this.state.gender);
 	}
 	// creates the function to apply the outro animation to the restaurant list when the restaurant details are expanded
 	listHide(){
@@ -84,7 +72,6 @@ class PostReview extends Component {
 	}
 	renderReview(){
 		var rating = parseInt(this.state.rating);	
-		// to convert date from : http://stackoverflow.com/questions/13459866/javascript-change-date-into-format-of-dd-mm-yyyy
 		var date = function js_yyyy_mm_dd_hh_mm_ss () {
 			  var now = new Date();
 			  var year = "" + now.getFullYear();
@@ -93,7 +80,7 @@ class PostReview extends Component {
 			  var hour = "" + now.getHours(); if (hour.length == 1) { hour = "0" + hour; }
 			  var minute = "" + now.getMinutes(); if (minute.length == 1) { minute = "0" + minute; }
 			  var second = "" + now.getSeconds(); if (second.length == 1) { second = "0" + second; }
-			  return year + "-" + month + "-" + day + " " + hour + ":" + minute + ":" + second;
+			  return year + "-" + month + "-" + day; //commenting out hour format for now " + hour + ":" + minute + ":" + second;
 			}();
 
 		return (
@@ -169,35 +156,11 @@ class PostReview extends Component {
 
 };
 
-// creates the function to join the action creator with the BookList component, to update the app's state
-// anything returned on this function, will end up as .props on the BookList container
 function mapDispatchToProps(dispatch) {
-
-	// Whenever selectBook is called, the result should be passed to all of our reducers
-	// returns the react function, joining selectBook (the action Creator) to the key : selectBook
 	return bindActionCreators({ 
-
-		// this is the KEY the function selectBook is bound to
-		// thanks to this key, the property from the action creator gets passed on to this.book
 		postReview: 
-
-		// this is the actual selectBook function
 		postReview },
-
-		// this is the argument
-		// 
-		// this actually BINDS the arguments above (selectBook) to all our reducers. If this function is called, 
-		// its passed to all the reducers, changing their state if needed
 		dispatch)
-
 }
 
-// this connects the two functions in this container together when exported
-// connect takes a function, and a component (class only), and produces a container
-// a container is again a component that connects react with the redux state
-// 
-// binds the original reducer state change and action creators to the BookList component
-// 
-// essentially this promotes BookList from a component to a container - react needs to know 
-// about this new component selection method, selectBook.
 export default connect(null, mapDispatchToProps) (PostReview);